bouge toutes les fixtures des tests dans shared
