+++ database/seeds/dummy/006-samples.ts
@@ -11,7 +11,6 @@ import { Region, Regions } from '../../../shared/referential/Region';
 import { Sample } from '../../../shared/schema/Sample/Sample';
 import { genCreatedSample } from '../../../shared/test/sampleFixtures';
 import { oneOf } from '../../../shared/test/testFixtures';
-import { Sampler1Fixture } from '../../../server/test/seed/001-users';
 import { DummyLaboratoryIds } from './002-laboratories';
 import {
   abricotsEtSimilaires,
@@ -22,6 +21,7 @@ import {
   lentilles,
   oignons
 } from './004-prescriptions';
+import { Sampler1Fixture } from '../../../shared/test/userFixtures';
 
 export const seed = async function () {
