+++ server/routers/test/auth.router.test.ts
@@ -1,6 +1,6 @@
 import { constants } from 'http2';
 import request from 'supertest';
-import { genUser } from '../../../shared/test/userFixtures';
+import { genUser, Sampler1Fixture } from '../../../shared/test/userFixtures';
 import { Users } from '../../repositories/userRepository';
 import { createServer } from '../../server';
 
@@ -8,7 +8,6 @@ import randomstring from 'randomstring';
 import { afterAll, beforeAll, describe, expect, test } from 'vitest';
 import { genAuthRedirectUrl } from '../../../shared/test/authFixtures';
 import { kysely } from '../../repositories/kysely';
-import { Sampler1Fixture } from '../../test/seed/001-users';
 import {
   mockAuthenticate,
   mockGetAuthorizationUrl,
