+++ server/routers/test/sample.router.test.ts
@@ -4,24 +4,12 @@ import  { omit } from 'lodash-es';
 import randomstring from 'randomstring';
 import request from 'supertest';
 import { v4 as uuidv4 } from 'uuid';
-import {
-  NationalCoordinator,
-  Sampler1Fixture,
-  Sampler2Fixture,
-} from '../../test/seed/001-users';
-import { ValidatedProgrammingPlanFixture } from '../../test/seed/002-programming-plans';
-import {
-  Sample11Fixture,
-  Sample12Fixture,
-  Sample13Fixture,
-  Sample2Fixture,
-} from '../../test/seed/004-samples';
 import { MatrixList } from '../../../shared/referential/Matrix/Matrix';
 import { Region, Regions } from '../../../shared/referential/Region';
 import {
   genCreatedPartialSample,
   genSampleContextData,
-  genSampleItem,
+  genSampleItem, Sample11Fixture, Sample12Fixture, Sample13Fixture, Sample2Fixture
 } from '../../../shared/test/sampleFixtures';
 import { oneOf } from '../../../shared/test/testFixtures';
 import { SampleItems } from '../../repositories/sampleItemRepository';
@@ -30,6 +18,8 @@ import { createServer } from '../../server';
 import { tokenProvider } from '../../test/testUtils';
 
 import { describe, test, expect } from 'vitest';
+import { NationalCoordinator, Sampler1Fixture, Sampler2Fixture } from '../../../shared/test/userFixtures';
+import { ValidatedProgrammingPlanFixture } from '../../../shared/test/programmingPlanFixtures';
 describe('Sample router', () => {
   const { app } = createServer();
