+++ server/test/seed/002-programming-plans.ts
@@ -1,16 +1,6 @@
 import { ProgrammingPlans } from '../../repositories/programmingPlanRepository';
-import { genProgrammingPlan } from '../../../shared/test/programmingPlanFixtures';
-import { NationalCoordinator } from './001-users';
+import { ValidatedProgrammingPlanFixture } from '../../../shared/test/programmingPlanFixtures';
 
-export const ValidatedProgrammingPlanFixture = genProgrammingPlan({
-  createdBy: NationalCoordinator.id,
-  createdAt: new Date('2024-01-01'),
-  id: '11111111-1111-1111-1111-111111111111',
-  status: 'Validated',
-  year: 2024,
-});
-
-export const seed = async () : Promise<void> => {
-
+export const seed = async (): Promise<void> => {
   await ProgrammingPlans().insert(ValidatedProgrammingPlanFixture);
 };
