+++ server/test/seed/003-companies.ts
@@ -1,15 +1,6 @@
 import { Companies } from '../../repositories/companyRepository';
-import { genCompany } from '../../../shared/test/companyFixtures';
+import { CompanyFixture } from '../../../shared/test/companyFixtures';
 
-export const CompanyFixture = genCompany({
-  siret: '11111111111111',
-  name: 'Company 1',
-  address: '1 rue de la Paix',
-  postalCode: '75000',
-});
-
-
-export const seed = async () : Promise<void> => {
-
+export const seed = async (): Promise<void> => {
   await Companies().insert(CompanyFixture);
-}
+};
