+++ server/test/seed/005-laboratories.ts
@@ -1,12 +1,6 @@
 import { Laboratories } from '../../repositories/laboratoryRepository';
-import { genLaboratory } from '../../../shared/test/laboratoryFixtures';
+import { LaboratoryFixture } from '../../../shared/test/laboratoryFixtures';
 
-export const LaboratoryFixture = genLaboratory({
-  id: '11111111-1111-1111-1111-111111111111',
-});
-
-
-export const seed = async () : Promise<void> => {
-
+export const seed = async (): Promise<void> => {
   await Laboratories().insert([LaboratoryFixture]);
 };
